horizontal_v1/codex/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0
pydantic-settings>=2.2.0
httpx>=0.27.0
anyio>=4.0.0
pillow>=10.0.0
PyJWT>=2.8.0
passlib>=1.7.4
python-multipart>=0.0.9
aiosqlite>=0.20.0
